pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "terrain-defi-protocol"
version = "0.1.0"
description = "DeFi Lending Protocol using 3D Terrain NFTs as Collateral"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
  { name = "Terrain Protocol Team" }
]

keywords = [
  "defi",
  "nft",
  "lending",
  "vyper",
  "uniswap",
  "dao",
  "blockchain"
]

dependencies = [
  "web3>=6.10.0",
  "python-dotenv>=1.0.0",
  "eth-abi>=5.0.0",
  "requests>=2.31.0",
  "pytest>=7.4.0"
]


[project.optional-dependencies]
dev = [
  "black>=24.0.0",
  "ruff>=0.1.0",
  "mypy>=1.6.0",
  "pytest-cov>=4.1.0",
]

analysis = [
  "numpy>=1.26.0",
  "pandas>=2.1.0",
]


[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]


[tool.black]
line-length = 88
target-version = ["py310"]
exclude = '''
/(
    \.venv
  | venv
  | env
  | build
  | dist
  | artifacts
)/
'''


[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]
exclude = [
  ".venv",
  "venv",
  "env",
  "build",
  "dist",
  "artifacts"
]


[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true
disallow_untyped_defs = false
exclude = "(build|dist|artifacts|venv)"


[tool.setuptools]
packages = ["bots", "scripts", "tests"]